server/requirements.txt
flask
flask_cors
python-dotenv
openai